Move data fetching of daily entry
