+++ pages/write.js
@@ -1,32 +1,13 @@
-import { graphql, useLazyLoadQuery } from 'react-relay/hooks';
-import { useStoreState } from 'easy-peasy';
-import { startOfDay } from 'date-fns';
+import { Suspense } from 'react';
 
 import withLayout from '../components/Layout';
 
 import Editor from '../components/Editor';
 
-const GET_ENTRY = graphql`
-  query writeQuery($userID: ID!, $date: String!) {
-    dailyEntry(userID: $userID, date: $date) {
-      id
-      content
-      wordCount
-      createdAt
-    }
-  }
-`;
-
-const Write = () => {
-  const { uid: userID } = useStoreState(state => state.user).firebaseData;
-
-  const date = startOfDay(new Date());
-
-  const { dailyEntry } = useLazyLoadQuery(GET_ENTRY, { userID, date });
-
-  return (
-    <Editor entry={dailyEntry} date={date} />
-  );
-}
+const Write = () => (
+  <Suspense fallback={<div>Loading...</div>}>
+    <Editor />
+  </Suspense>
+);
 
 export default withLayout(Write);
